FIX: publish error
