+++ include/cv_camera/capture.hpp
@@ -7,7 +7,6 @@
 #include <string>
 #include <thread>
 #include "cv_camera/exception.h"
-#include "cv_camera/utils.hpp"
 
 #include <cv_bridge/cv_bridge.h>
 #include <camera_info_manager/camera_info_manager.hpp>
@@ -172,6 +171,37 @@ class Capture
      * @brief rescale camera calibration to another resolution
      */
     void rescaleCameraInfo(uint width, uint height);
+
+    std::string mat_type2encoding(int mat_type)
+    {
+        switch (mat_type)
+        {
+            case CV_8UC1:
+                return "mono8";
+            case CV_8UC3:
+                return "bgr8";
+            case CV_16SC1:
+                return "mono16";
+            case CV_8UC4:
+                return "rgba8";
+            default:
+                throw std::runtime_error("Unsupported encoding type");
+        }
+    }
+
+    void set_now(builtin_interfaces::msg::Time& time)
+    {
+        std::chrono::nanoseconds now = std::chrono::high_resolution_clock::now().time_since_epoch();
+        if (now <= std::chrono::nanoseconds(0))
+        {
+            time.sec = time.nanosec = 0;
+        }
+        else
+        {
+            time.sec = static_cast<builtin_interfaces::msg::Time::_sec_type>(now.count() / 1000000000);
+            time.nanosec = now.count() % 1000000000;
+        }
+    }
 
     /**
      * @brief node handle for advertise.
